Move normalize.ts into its own folder
